Updated Components
Position Component - Changed ints to floats
Render Component - Changed assorted things and added member functions.

+++ PositionComponent.h
@@ -28,7 +28,7 @@ public:
     {
         return _Position;
     }
-    void SetPosition(int x, int y)
+    void SetPosition(float x, float y)
     {
         _Position.x = x;
         _Position.y = y;
